// api-gateway/src/common/enums/posts.command.enums.ts
export enum PostsCommand {
  NEW_POST = 'NEW_POST',
  EDIT_POST = 'EDIT_POST',
  DELETE_POST = 'DELETE_POST',
  GET_POST = 'GET_POST',
  GET_POSTS = 'GET_POSTS',
  ADD_COMMENT = 'ADD_COMMENT',
  DELETE_COMMENT = 'DELETE_COMMENT',
  EDIT_COMMENT = 'EDIT_COMMENT',
  LIKE_COMMENT = 'LIKE_COMMENT',
  LIKE_POST = 'LIKE_POST',
  GET_LIKE_LIST_POST = 'GET_LIKE_LIST_POST',
  GET_LIKE_LIST_COMMENT = 'GET_LIKE_LIST_COMMENT',
  SEND_ARRAY_USER_ID = 'SEND_ARRAY_USER_ID'
}